chore(alert): add a pseudo-localization story
