+++ package.json
@@ -197,6 +197,7 @@
     "precise-commits": "^1.0.2",
     "prettier": "^2.2.1",
     "prop-types": "^15.7.2",
+    "pseudo-localization": "2.1.1",
     "puppeteer-core": "^10.1.0",
     "react": "^17.0.2",
     "react-16": "npm:react@^16.8.6",
